\name{boot.compute.fi.fj}
\alias{boot.compute.fi.fj}
\title{
Bootstrap computations for TInPosition.
}
\description{

}
\usage{
boot.compute.fi.fj(DATA, DESIGN, res)
}

\arguments{
  \item{DATA}{

}
  \item{DESIGN}{

}
  \item{res}{

}
}
\details{

}
\value{

}
\references{

}
\author{
Derek Beaton
}

\seealso{

}
\examples{

}
\keyword{ bootstrap }
