\name{print.tepBADA.inference.battery}
\alias{print.tepBADA.inference.battery}
\title{Print tepBADA inference results}
\description{
	Print tepBADA inference results.
}
\usage{
	\method{print}{tepBADA.inference.battery}(x,\dots)
}
\arguments{
  \item{x}{an list that contains items to make into the tepBADA.inference.battery class.}
  \item{\dots}{inherited/passed arguments for S3 print method(s).}  
 }
\author{Derek Beaton, Cherise Chin-Fatt}
\keyword{print}
