% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tepDICA.inference.battery.R
\name{print.tepDICA.inference.battery}
\alias{print.tepDICA.inference.battery}
\title{Print tepDICA.inference.battery results}
\usage{
\method{print}{tepDICA.inference.battery}(x, ...)
}
\arguments{
\item{x}{an list that contains items to make into the
tepDICA.inference.battery class.}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print tepDICA Inference results.
}
\author{
Derek Beaton, Cherise Chin-Fatt
}
\keyword{print}
