% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tinpoBoot.R
\name{print.tinpoBoot}
\alias{print.tinpoBoot}
\title{Print results from TInPosition Bootstraps}
\usage{
\method{print}{tinpoBoot}(x, ...)
}
\arguments{
\item{x}{an list that contains items to make into the tinpoBoot class.}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print bootstrap results from the TInPosition.
}
\author{
Derek Beaton and Cherise Chin-Fatt
}
\keyword{print}
