% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tinpoBootTests.R
\name{print.tinpoBootTests}
\alias{print.tinpoBootTests}
\title{Print results from TInPosition Bootstrap Ratio Tests}
\usage{
\method{print}{tinpoBootTests}(x, ...)
}
\arguments{
\item{x}{an list that contains items to make into the tinpoBootTests class.}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print bootstrap ratio tests results from the TInPosition.
}
\author{
Derek Beaton and Cherise Chin-Fatt
}
\keyword{print}
