\name{print.tinpoComponents}
\alias{print.tinpoComponents}
\title{Print results from TInPosition Components Permutation Test}
\description{
	Print Components permutation test results from the TInPosition.
}
\usage{
	\method{print}{tinpoComponents}(x,\dots)
}
\arguments{
  \item{x}{an list that contains items to make into the tinpoComponents class.}
  \item{\dots}{inherited/passed arguments for S3 print method(s).}  
 }
\author{Derek Beaton and Cherise Chin-Fatt}
\keyword{print}
