% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tinpoLOO.R
\name{print.tinpoLOO}
\alias{print.tinpoLOO}
\title{Print results from TInPosition LOO}
\usage{
\method{print}{tinpoLOO}(x, ...)
}
\arguments{
\item{x}{an list that contains items to make into the tinpoLOO class.}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print LOO results from the TInPosition.
}
\author{
Derek Beaton and Cherise Chin-Fatt
}
\keyword{print}
