% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tinpoOmni.R
\name{print.tinpoOmni}
\alias{print.tinpoOmni}
\title{Print results from TInPosition Omnibus Permutation Test}
\usage{
\method{print}{tinpoOmni}(x, ...)
}
\arguments{
\item{x}{an list that contains items to make into the tinpoOmni class.}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print Omnibus permutation test results from the TInPosition.
}
\author{
Derek Beaton and Cherise Chin-Fatt
}
\keyword{print}
