\name{print.tinpoOutput}
\alias{print.tinpoOutput}
\title{Print results from TInPosition}
\description{
	Print results from the TInPosition.
}
\usage{
	\method{print}{tinpoOutput}(x,\dots)
}
\arguments{
  \item{x}{an list that contains items to make into the tinpoOutput class.}
  \item{\dots}{inherited/passed arguments for S3 print method(s).}  
 }
\seealso{
	\code{\link{tepBADA.inference.battery}}, \code{\link{tinGraphs}}
}
\author{Derek Beaton and Cherise Chin-Fatt}
\keyword{print}
