% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tinpoR2.R
\name{print.tinpoR2}
\alias{print.tinpoR2}
\title{Print results from TInPosition R2 Permutation Test}
\usage{
\method{print}{tinpoR2}(x, ...)
}
\arguments{
\item{x}{an list that contains items to make into the tinpoR2 class.}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print R2 permutation test results from the TInPosition.
}
\author{
Derek Beaton and Cherise Chin-Fatt
}
\keyword{print}
