\name{TInPosition-package}
\alias{TInPosition-package}
\alias{TInPosition}
\encoding{UTF-8}
\docType{package}
\title{
TInPosition: Inference Tests for Two-table \emph{ExPosition} (\code{TExPosition}).
}
\description{
TInPosition provides multiple forms of inference tests for the \code{\link{TExPosition}} package. 
}
\details{
\tabular{ll}{
Package: \tab TInPosition\cr
Type: \tab Package\cr
Version: \tab 0.13.8\cr
Date: \tab 2020-11-19\cr
Depends: \tab R (>=2.15.0), prettyGraphs (>= 2.1.4), TExPosition (>= 2.6.10), ExPosition (>= 2.8.19), InPosition (>= 0.12.7)\cr
License: \tab GPL-2\cr
URL: \tab https://github.com/derekbeaton/ExPosition1\cr
}
}
\author{
Questions, comments, compliments, and complaints go to Derek Beaton \email{exposition.software@gmail.com}. Also see the bug-tracking and live update website for ExPosition: https://github.com/derekbeaton/ExPosition1/\cr\cr 
Primary authors and contributors are: Derek Beaton, Jenny Rieck, and Hervé Abdi
}
\references{
	Permutation:\cr
	Berry, K. J., Johnston, J. E., & Mielke, P. W. (2011). Permutation methods. \emph{Wiley Interdisciplinary Reviews: Computational Statistics},\emph{3}, 527--542. \cr
	Peres-Neto, P. R., Jackson, D. A., & Somers, K. M. (2005). How many principal components? Stopping rules for determining the number of non-trivial axes revisited. \emph{Computational Statistics & Data Analysis}, \emph{49}(\bold{4}), 974--997.\cr\cr
	Bootstrap:\cr
	Chernick, M. R. (2008). \emph{Bootstrap methods: A guide for practitioners and researchers} (Vol. 619). Wiley-Interscience.\cr
	Hesterberg, T. (2011). Bootstrap. \emph{Wiley Interdisciplinary Reviews: Computational Statistics}, \emph{3}, 497--526. \cr\cr
	Two-table specific cases:\cr
	Krishnan, A., Williams, L. J., McIntosh, A. R., & Abdi, H. (2011). Partial Least Squares (PLS) methods for neuroimaging: A tutorial and review. \emph{NeuroImage}, \emph{56}(\bold{2}), 455 -- 475.\cr
	McIntosh, A. R., & Lobaugh, N. J. (2004). Partial least squares analysis of neuroimaging data: applications and advances. \emph{Neuroimage}, \emph{23}, S250--S263.\cr
	Abdi, H., Williams, L.J., Beaton, D., Posamentier, M., Harris, T.S., Krishnan, A., & Devous, M.D. (2012). Analysis of regional cerebral blood flow data to discriminate among Alzheimer's disease, fronto-temporal dementia, and elderly controls: A multi-block barycentric discriminant analysis (MUBADA) methodology. \emph{Journal of Alzheimer Disease}, \bold{31}, s189--s201.\cr
}
\keyword{ package }
\keyword{ multivariate }
\seealso{
\code{\link{tepBADA.inference.battery}}, \code{\link{tepDICA.inference.battery}}\cr\cr
}
\examples{
#For more examples, see each individual function (as noted above).
}
