% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.compute.fi.fj.R
\name{boot.compute.fi.fj}
\alias{boot.compute.fi.fj}
\title{Bootstrap computations for TInPosition.}
\usage{
boot.compute.fi.fj(DATA, DESIGN, res)
}
\arguments{
\item{DATA}{The original data matrix to be bootstrapped. Rows will be
bootstrapped and are assumed to be observations. Resampling will be
constrained to within groups based on \code{DESIGN}.}

\item{DESIGN}{A design matrix (in disjunctive coding). Required for
\code{TExPosition} and \code{TInPosition} analyses.}

\item{res}{of class \code{texpoOutput}. Results from one of the
\code{\link[TExPosition]{TExPosition}} methods (e.g., \code{\link[TExPosition]{tepDICA}},
\code{\link[TExPosition]{tepBADA}}),}
}
\value{
\item{FBX}{a set of factor scores of the measures (columns,
\code{$fj}) for the bootstrapped data.} \item{FBY}{a set of factor scores of
the groups (\code{$fi}) for the bootstrapped data.}
}
\description{
Provides bootstrap projections for $fi and $fj from TExPosition methods.
}
\author{
Derek Beaton
}
\keyword{bootstrap}
