% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tepBADA.inference.battery.R
\name{print.tepBADA.inference.battery}
\alias{print.tepBADA.inference.battery}
\title{Print tepBADA inference results}
\usage{
\method{print}{tepBADA.inference.battery}(x, ...)
}
\arguments{
\item{x}{an list that contains items to make into the
tepBADA.inference.battery class.}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print tepBADA inference results.
}
\author{
Derek Beaton, Cherise Chin-Fatt
}
\keyword{print}
