\name{print.tepDICA.inference.battery}
\alias{print.tepDICA.inference.battery}
\title{Print tepDICA.inference.battery results}
\description{
	Print tepDICA Inference results.
}
\usage{
	\method{print}{tepDICA.inference.battery}(x,\dots)
}
\arguments{
  \item{x}{an list that contains items to make into the tepDICA.inference.battery class.}
  \item{\dots}{inherited/passed arguments for S3 print method(s).}  
 }
\author{Derek Beaton, Cherise Chin-Fatt}
\keyword{print}
