\name{print.tinpoBoot}
\alias{print.tinpoBoot}
\title{Print results from TInPosition Bootstraps}
\description{
	Print bootstrap results from the TInPosition.
}
\usage{
	\method{print}{tinpoBoot}(x,\dots)
}
\arguments{
  \item{x}{an list that contains items to make into the tinpoBoot class.}
  \item{\dots}{inherited/passed arguments for S3 print method(s).}  
 }
\author{Derek Beaton and Cherise Chin-Fatt}
\keyword{print}
