\name{print.tinpoBootTests}
\alias{print.tinpoBootTests}
\title{Print results from TInPosition Bootstrap Ratio Tests}
\description{
	Print bootstrap ratio tests results from the TInPosition.
}
\usage{
	\method{print}{tinpoBootTests}(x,\dots)
}
\arguments{
  \item{x}{an list that contains items to make into the tinpoBootTests class.}
  \item{\dots}{inherited/passed arguments for S3 print method(s).}  
 }
\author{Derek Beaton and Cherise Chin-Fatt}
\keyword{print}
