% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tinpoComponents.R
\name{print.tinpoComponents}
\alias{print.tinpoComponents}
\title{Print results from TInPosition Components Permutation Test}
\usage{
\method{print}{tinpoComponents}(x, ...)
}
\arguments{
\item{x}{an list that contains items to make into the tinpoComponents class.}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print Components permutation test results from the TInPosition.
}
\author{
Derek Beaton and Cherise Chin-Fatt
}
\keyword{print}
