\name{print.tinpoLOO}
\alias{print.tinpoLOO}
\title{Print results from TInPosition LOO}
\description{
	Print LOO results from the TInPosition.
}
\usage{
	\method{print}{tinpoLOO}(x,\dots)
}
\arguments{
  \item{x}{an list that contains items to make into the tinpoLOO class.}
  \item{\dots}{inherited/passed arguments for S3 print method(s).}  
 }
\author{Derek Beaton and Cherise Chin-Fatt}
\keyword{print}
