\name{print.tinpoOmni}
\alias{print.tinpoOmni}
\title{Print results from TInPosition Omnibus Permutation Test}
\description{
	Print Omnibus permutation test results from the TInPosition.
}
\usage{
	\method{print}{tinpoOmni}(x,\dots)
}
\arguments{
  \item{x}{an list that contains items to make into the tinpoOmni class.}
  \item{\dots}{inherited/passed arguments for S3 print method(s).}  
 }
\author{Derek Beaton and Cherise Chin-Fatt}
\keyword{print}
