% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tinpoOutput.R
\name{print.tinpoOutput}
\alias{print.tinpoOutput}
\title{Print results from TInPosition}
\usage{
\method{print}{tinpoOutput}(x, ...)
}
\arguments{
\item{x}{an list that contains items to make into the tinpoOutput class.}

\item{\dots}{inherited/passed arguments for S3 print method(s).}
}
\description{
Print results from the TInPosition.
}
\seealso{
\code{\link{tepBADA.inference.battery}}, \code{\link{tinGraphs}}
}
\author{
Derek Beaton and Cherise Chin-Fatt
}
\keyword{print}
