\name{print.tinpoR2}
\alias{print.tinpoR2}
\title{Print results from TInPosition R2 Permutation Test}
\description{
	Print R2 permutation test results from the TInPosition.
}
\usage{
	\method{print}{tinpoR2}(x,\dots)
}
\arguments{
  \item{x}{an list that contains items to make into the tinpoR2 class.}
  \item{\dots}{inherited/passed arguments for S3 print method(s).}  
 }
\author{Derek Beaton and Cherise Chin-Fatt}
\keyword{print}
